+++ pysui/sui/sui_pgql/execute/caching_txn.py
@@ -14,16 +14,15 @@ from pysui.sui.sui_pgql.pgql_txb_signing import SignerBlock
 from pysui.sui.sui_pgql.pgql_txn_base import _TransactionBase as txbase
 
 from pysui.sui.sui_types import bcs
-import pysui.sui.sui_pgql.pgql_txb_gas as gd
 import pysui.sui.sui_pgql.pgql_validators as tv
 import pysui.sui.sui_pgql.pgql_query as qn
 import pysui.sui.sui_pgql.pgql_types as pgql_type
 import pysui.sui.sui_pgql.pgql_txn_async_argb as argbase
 from .caching_tx_builder import CachingTransactionBuilder, PureInput
-from pysui.sui.sui_common.async_lru import AsyncLRU
+
+from pysui.sui.sui_common.async_funcs import AsyncLRU
 
 logger = logging.getLogger("serial_exec")
-# logger = logging.getLogger("caching transaction")
 
 
 class CachingTransaction(txbase):